Module aws_sdk_apprunner::error

Modules

internal_service_error_exception

See InternalServiceErrorException

invalid_request_exception

See InvalidRequestException

invalid_state_exception

See InvalidStateException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

Structs

AssociateCustomDomainError

CreateAutoScalingConfigurationError

CreateConnectionError

CreateServiceError

DeleteAutoScalingConfigurationError

DeleteConnectionError

DeleteServiceError

DescribeAutoScalingConfigurationError

DescribeCustomDomainsError

DescribeServiceError

DisassociateCustomDomainError

InternalServiceErrorException

An unexpected service exception occurred.

InvalidRequestException

One or more input parameters aren't valid. Refer to the API action's document page, correct the input parameters, and try the action again.

InvalidStateException

You can't perform this action when the resource is in its current state.

ListAutoScalingConfigurationsError

ListConnectionsError

ListOperationsError

ListServicesError

ListTagsForResourceError

PauseServiceError

ResourceNotFoundException

A resource doesn't exist for the specified Amazon Resource Name (ARN) in your AWS account.

ResumeServiceError

ServiceQuotaExceededException

App Runner can't create this resource. You've reached your account quota for this resource type.

StartDeploymentError

TagResourceError

UntagResourceError

UpdateServiceError

Enums

AssociateCustomDomainErrorKind

CreateAutoScalingConfigurationErrorKind

CreateConnectionErrorKind

CreateServiceErrorKind

DeleteAutoScalingConfigurationErrorKind

DeleteConnectionErrorKind

DeleteServiceErrorKind

DescribeAutoScalingConfigurationErrorKind

DescribeCustomDomainsErrorKind

DescribeServiceErrorKind

DisassociateCustomDomainErrorKind

ListAutoScalingConfigurationsErrorKind

ListConnectionsErrorKind

ListOperationsErrorKind

ListServicesErrorKind

ListTagsForResourceErrorKind

PauseServiceErrorKind

ResumeServiceErrorKind

StartDeploymentErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateServiceErrorKind