Module aws_sdk_apprunner::model

Modules

authentication_configuration

See AuthenticationConfiguration

auto_scaling_configuration

See AutoScalingConfiguration

auto_scaling_configuration_summary

See AutoScalingConfigurationSummary

certificate_validation_record

See CertificateValidationRecord

code_configuration

See CodeConfiguration

code_configuration_values

See CodeConfigurationValues

code_repository

See CodeRepository

connection

See Connection

connection_summary

See ConnectionSummary

custom_domain

See CustomDomain

encryption_configuration

See EncryptionConfiguration

health_check_configuration

See HealthCheckConfiguration

image_configuration

See ImageConfiguration

image_repository

See ImageRepository

instance_configuration

See InstanceConfiguration

operation_summary

See OperationSummary

service

See Service

service_summary

See ServiceSummary

source_code_version

See SourceCodeVersion

source_configuration

See SourceConfiguration

tag

See Tag

Structs

AuthenticationConfiguration

Describes resources needed to authenticate access to some source repositories. The specific resource depends on the repository provider.

AutoScalingConfiguration

Describes an AWS App Runner automatic scaling configuration resource. Multiple revisions of a configuration have the same

AutoScalingConfigurationSummary

Provides summary information about an AWS App Runner automatic scaling configuration resource.

CertificateValidationRecord

Describes a certificate CNAME record to add to your DNS. For more information, see AssociateCustomDomain.

CodeConfiguration

Describes the configuration that AWS App Runner uses to build and run an App Runner service from a source code repository.

CodeConfigurationValues

Describes the basic configuration needed for building and running an AWS App Runner service. This type doesn't support the full set of possible

CodeRepository

Describes a source code repository.

Connection

Describes an AWS App Runner connection resource.

ConnectionSummary

Provides summary information about an AWS App Runner connection resource.

CustomDomain

Describes a custom domain that's associated with an AWS App Runner service.

EncryptionConfiguration

Describes a custom encryption key that AWS App Runner uses to encrypt copies of the source repository and service logs.

HealthCheckConfiguration

Describes the settings for the health check that AWS App Runner performs to monitor the health of a service.

ImageConfiguration

Describes the configuration that AWS App Runner uses to run an App Runner service using an image pulled from a source image repository.

ImageRepository

Describes a source image repository.

InstanceConfiguration

Describes the runtime configuration of an AWS App Runner service instance (scaling unit).

OperationSummary

Provides summary information for an operation that occurred on an AWS App Runner service.

Service

Describes an AWS App Runner service. It can describe a service in any state, including deleted services.

ServiceSummary

Provides summary information for an AWS App Runner service.

SourceCodeVersion

Identifies a version of code that AWS App Runner refers to within a source code repository.

SourceConfiguration

Describes the source deployed to an AWS App Runner service. It can be a code or an image repository.

Tag

Describes a tag that is applied to an AWS App Runner resource. A tag is a metadata item consisting of a key-value pair.

Enums

AutoScalingConfigurationStatus

CertificateValidationRecordStatus

ConfigurationSource

ConnectionStatus

CustomDomainAssociationStatus

HealthCheckProtocol

ImageRepositoryType

OperationStatus

OperationType

ProviderType

Runtime

ServiceStatus

SourceCodeVersionType