Struct aws_sdk_apprunner::model::service::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Service

Implementations

impl Builder

pub fn service_name(self, input: impl Into<String>) -> Self

The customer-provided service name.

pub fn set_service_name(self, input: Option<String>) -> Self

pub fn service_id(self, input: impl Into<String>) -> Self

An ID that App Runner generated for this service. It's unique within the AWS Region.

pub fn set_service_id(self, input: Option<String>) -> Self

pub fn service_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of this service.

pub fn set_service_arn(self, input: Option<String>) -> Self

pub fn service_url(self, input: impl Into<String>) -> Self

A subdomain URL that App Runner generated for this service. You can use this URL to access your service web application.

pub fn set_service_url(self, input: Option<String>) -> Self

pub fn created_at(self, input: Instant) -> Self

The time when the App Runner service was created. It's in the Unix time stamp format.

pub fn set_created_at(self, input: Option<Instant>) -> Self

pub fn updated_at(self, input: Instant) -> Self

The time when the App Runner service was last updated at. It's in the Unix time stamp format.

pub fn set_updated_at(self, input: Option<Instant>) -> Self

pub fn deleted_at(self, input: Instant) -> Self

The time when the App Runner service was deleted. It's in the Unix time stamp format.

pub fn set_deleted_at(self, input: Option<Instant>) -> Self

pub fn status(self, input: ServiceStatus) -> Self

The current state of the App Runner service. These particular values mean the following.

CREATE_FAILED – The service failed to create. To troubleshoot this failure, read the failure events and logs, change any parameters that need to be fixed, and retry the call to create the service.

The failed service isn't usable, and still counts towards your service quota. When you're done analyzing the failure, delete the service.

DELETE_FAILED – The service failed to delete and can't be successfully recovered. Retry the service deletion call to ensure that all related resources are removed.

pub fn set_status(self, input: Option<ServiceStatus>) -> Self

pub fn source_configuration(self, input: SourceConfiguration) -> Self

The source deployed to the App Runner service. It can be a code or an image repository.

pub fn set_source_configuration(
self,
input: Option<SourceConfiguration>
) -> Self

pub fn instance_configuration(self, input: InstanceConfiguration) -> Self

The runtime configuration of instances (scaling units) of this service.

pub fn set_instance_configuration(
self,
input: Option<InstanceConfiguration>
) -> Self

pub fn encryption_configuration(self, input: EncryptionConfiguration) -> Self

The encryption key that App Runner uses to encrypt the service logs and the copy of the source repository that App Runner maintains for the service. It can be either a customer-provided encryption key or an AWS managed CMK.

pub fn set_encryption_configuration(
self,
input: Option<EncryptionConfiguration>
) -> Self

pub fn health_check_configuration(self, input: HealthCheckConfiguration) -> Self

The settings for the health check that App Runner performs to monitor the health of this service.

pub fn set_health_check_configuration(
self,
input: Option<HealthCheckConfiguration>
) -> Self

pub fn auto_scaling_configuration_summary(
self,
input: AutoScalingConfigurationSummary
) -> Self

Summary information for the App Runner automatic scaling configuration resource that's associated with this service.

pub fn set_auto_scaling_configuration_summary(
self,
input: Option<AutoScalingConfigurationSummary>
) -> Self

pub fn build(self) -> Service

Consumes the builder and constructs a Service