Module aws_sdk_apprunner::operation

Structs

AssociateCustomDomain

Associate your own domain name with the AWS App Runner subdomain URL of your App Runner service.

CreateAutoScalingConfiguration

Create an AWS App Runner automatic scaling configuration resource. App Runner requires this resource

CreateConnection

Create an AWS App Runner connection resource. App Runner requires a connection resource when you create App Runner services that access private repositories from

CreateService

Create an AWS App Runner service. After the service is created, the action also automatically starts a deployment.

DeleteAutoScalingConfiguration

Delete an AWS App Runner automatic scaling configuration resource. You can delete a specific revision or the latest active revision. You can't delete a

DeleteConnection

Delete an AWS App Runner connection. You must first ensure that there are no running App Runner services that use this connection. If there are any, the

DeleteService

Delete an AWS App Runner service.

DescribeAutoScalingConfiguration

Return a full description of an AWS App Runner automatic scaling configuration resource.

DescribeCustomDomains

Return a description of custom domain names that are associated with an AWS App Runner service.

DescribeService

Return a full description of an AWS App Runner service.

DisassociateCustomDomain

Disassociate a custom domain name from an AWS App Runner service.

ListAutoScalingConfigurations

Returns a list of AWS App Runner automatic scaling configurations in your AWS account. You can query the revisions for a specific configuration name or

ListConnections

Returns a list of AWS App Runner connections that are associated with your AWS account.

ListOperations

Return a list of operations that occurred on an AWS App Runner service.

ListServices

Returns a list of running AWS App Runner services in your AWS account.

ListTagsForResource

List tags that are associated with for an AWS App Runner resource. The response contains a list of tag key-value pairs.

PauseService

Pause an active AWS App Runner service. App Runner reduces compute capacity for the service to zero and loses state (for example, ephemeral storage is

ResumeService

Resume an active AWS App Runner service. App Runner provisions compute capacity for the service.

StartDeployment

Initiate a manual deployment of the latest commit in a source code repository or the latest image in a source image repository to an AWS App Runner

TagResource

Add tags to, or update the tag values of, an App Runner resource. A tag is a key-value pair.

UntagResource

Remove tags from an App Runner resource.

UpdateService

Update an AWS App Runner service. You can update the source configuration and instance configuration of the service. You can also update the ARN of the auto